+++ src/eventsub.rs
@@ -1,3 +1,5 @@
 /// An EventSub Subscription ID
 #[aliri_braid::braid(serde)]
 pub struct EventSubId;
+
+impl_extra!(EventSubId, EventSubIdRef);